refactor(hooks): extract useTabSync from use-editor-groups
